Extract search route handling into composable, set Logo loading when fetching
