+++ src/components/VHeader/VHeader.vue
@@ -4,7 +4,7 @@
     :class="{ 'border-b border-dark-charcoal-20': isHeaderScrolled }"
   >
     <NuxtLink to="/">
-      <VLogoLoader />
+      <VLogoLoader :status="isFetching ? 'loading' : 'idle'" />
     </NuxtLink>
     <VButton
       v-if="!!currentOverlay"
@@ -23,20 +23,17 @@ import {
   computed,
   defineComponent,
   ref,
-  useRoute,
-  useRouter,
+  useContext,
 } from '@nuxtjs/composition-api'
+
+import { isScreen } from '~/composables/use-media-query'
+import { useSearchRoute } from '~/composables/use-search-route'
 import { useWindowScroll } from '~/composables/use-window-scroll'
 
 import closeIcon from '~/assets/icons/close.svg'
 
-import VLogoLoader from '~/components/VLogoLoader/VLogoLoader.vue'
 import VIcon from '~/components/VIcon/VIcon.vue'
-import { isScreen } from '~/composables/use-media-query'
-
-const HEADER_HEIGHT = 80
-
-const searchRoutes = ['search', 'search-image', 'search-audio', 'search-video']
+import VLogoLoader from '~/components/VLogoLoader/VLogoLoader.vue'
 
 const VHeader = defineComponent({
   name: 'VHeader',
@@ -45,31 +42,37 @@ const VHeader = defineComponent({
     VLogoLoader,
   },
   setup() {
-    const { y } = useWindowScroll()
-    const isHeaderScrolled = computed(() => y.value > HEADER_HEIGHT)
-
+    const { store } = useContext()
+    const { isSearch } = useSearchRoute()
+    const { isHeaderScrolled } = useWindowScroll()
     const isMdScreen = isScreen('md')
+
+    /** @type {import('@nuxtjs/composition-api').Ref<null|'filters'|'content-switcher'>} */
     const currentOverlay = ref(null)
+    /**
+     * When an overlay is opened on mobile, this sets the current overlay name
+     * @param {'filters'|'content-switcher'} overlay
+     */
     // eslint-disable-next-line no-unused-vars
-    const setCurrentOverlay = () => {
+    const setCurrentOverlay = (overlay) => {
       // Overlay can only be set on mobile screen
-      if (!isMdScreen.value) return
+      if (isMdScreen.value) return
+      currentOverlay.value = overlay
     }
     const closeOverlay = () => {
       currentOverlay.value = null
     }
-    const route = useRoute()
-    const router = useRouter()
-    const isSearch = ref(route.value.name === 'search')
-    router.beforeEach((to, from, next) => {
-      isSearch.value = searchRoutes.includes(to.name)
-      next()
-    })
+
+    /**  @type {import('@nuxtjs/composition-api').ComputedRef<Boolean>} */
+    const isFetching = computed(
+      () => store.getters['media/fetchingState'].isFetching
+    )
 
     return {
       closeIcon,
       closeOverlay,
       currentOverlay,
+      isFetching,
       isHeaderScrolled,
       isSearch,
     }
