+++ src/composables/use-window-scroll.js
@@ -4,6 +4,15 @@ import { defaultWindow } from '~/composables/window'
 import { ref } from '@nuxtjs/composition-api'
 import { useEventListener } from '~/composables/use-event-listener'
 
+const HEADER_HEIGHT = 80
+
+/**
+ * Whether the page has been scrolled down past the header.
+ *
+ * @type {import('@nuxtjs/composition-api').Ref<boolean>}
+ */
+export const isHeaderScrolled = ref(false)
+
 /**
  *
  * @param {object} options
@@ -28,6 +37,7 @@ export function useWindowScroll({
   const scrollHandler = () => {
     x.value = window.pageXOffset
     y.value = window.pageYOffset
+    isHeaderScrolled.value = y.value > HEADER_HEIGHT
   }
 
   const handler = throttleMs
@@ -39,5 +49,5 @@ export function useWindowScroll({
     passive: true,
   })
 
-  return { x, y }
+  return { x, y, isHeaderScrolled }
 }
